test: Move fixture work_dir, accessible to all tests
