+++ src/tests/conftest.py
@@ -1,0 +1,1 @@
+from .fixtures.work_dir import work_dir
